inline system_api_security_test
