+++ rs/tests/src/execution/mod.rs
@@ -1,5 +1,3 @@
-pub mod system_api_security_test;
-
 use ic_registry_subnet_type::SubnetType;
 use ic_system_test_driver::driver::{
     ic::{InternetComputer, Subnet},
